Move speed test device implementations into gateware library.
